+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The failure is due to the storage manager which is already loaded and a new inst
 * Update the design of context management. Add a new API `smw_allocate_context()` to allocate context before initializing multi-part operation. Automatically release context resources if final multi-part operation succeeds, or upon encountering any critical error during the multi-part cryptographic operation or by calling the `smw_cancel_operation()` API.
 * Update the arguments passed to `smw_cancel_operation()` and `smw_copy_context()`.
 * Support storing the IV used by subsystem in the `output_iv` field for the AEAD final encryption operation, in addition to existing support for the one-shot AEAD encryption operation.
+* Fix the TLV lifecycle encoding to handle boolean. Make TLV lifecycle as core global function.
 
 ##### 2. Subsystems